% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_median.R
\name{mean_median}
\alias{mean_median}
\title{mean_median}
\usage{
mean_median(x)
}
\arguments{
\item{x}{is greater than or equal to 2 numeric variables.}
}
\value{
a vector containing the mean and median of the variables.
}
\description{
This function calculates the mean and median of a group of numeric variables.
}
\examples{
mean_median(c(2,4,6,8,2,3,10,5))
}
