\name{MeanMedian-package}
\alias{MeanMedian-package}
\alias{MeanMedian}
\docType{package}
\title{
\packageTitle{MeanMedian}
}
\description{
\packageDescription{MeanMedian}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{MeanMedian}

Maintainer: \packageMaintainer{MeanMedian}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = MeanMedian)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{MeanMedian}}
%\section{Documentation Index}{\packageIndices{MeanMedian}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
