% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_median.R
\name{mean_median}
\alias{mean_median}
\title{Calculate mean minus median}
\usage{
mean_median(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A number: the mean(x) minus median(x)
}
\description{
Calculate mean minus median
}
\examples{
mean_median(c(1,2,3))
}
